Refactor make_complete() of MultiIndexSet
The flag parameter `inplace` has been added to the `make_complete()`
method of the `MultiIndexSet` class. When the parameter is set to
`True`, the modification to underlying exponents will be made
in-place; when `False` (the default), a new instance with a complete
exponents will be created. If the set is already complete,
a shallow copy will be created.

This commit should resolve Issue #103.