Refactor MultiIndexSet.add_exponents() with inplace
The method `add_exponents()` of the `MultiIndexSet` class
has been updated with an `inplace` flag parameter to allow
the modification to the underlying exponents to happen either
directly in the current instance or
as exponents of a new instance (the default behavior).
This commit should resolve Issue #117.

A couple of additional changes:

- The test suite has been extended to test the behavior of adding
  exponents both inplace and outplace.
- The in-code documentation of `add_exponents()` has been updated
  to conform with the Minterpy standards.
- The function `core.verification.check_values()` has been refactored
  with improved in-code documentation.

Because the hidden method `_new_instance_if_necessary()` is not used
anymore by `add_exponents()`, Issue #81 is also resolved by this commit.
The test suite has been extended to check the previously broken behavior.