Refactor MultiIndexSet.expand_dim() with inplace
The instance method `expand_dim()` of the `MultiIndexSet` class
has been updated with an `inplace` flag parameter (thus conforming with
Issues #103 and #117) to allow the modification to the underlying exponents
to happen either directly in the current instance or as exponents of a new
instance (this is now the default behavior).
This MR should resolve Issue #118.

Several additional changes:

- Because the default behavior of calling the method is now changed
  (outplace instead inplace), the method `expand_dim()` of the
  `MultivariatePolynomialSingleABC` class has also been updated such that
  the modification to the multi-index set happens outplace.
- The test suite has been extended to test the behavior of expanding
  the dimension of an exponent array.
- The in-code documentation of `expand_dim()` has been updated to conform
  with the Minterpy standards.
- The function `core.utils.expand_dim()` has been refactored and
  its in-code documentation updated.
- The test suite now also includes tests related to calling the utility
  function `core.utils.expand_dim()` which previously was not tested directly.