Spatial dimension check in `MultiIndexSet.from_degree()` is more forgiving.
The type and value check for spatial dimension in the constructor
`MultiIndexSet.from_degree()` is now more forgiving as long as
the provided value:

- is strictly positive
- is a whole number
- can be compared (`<= 0`) in an `if` statement without ambiguity
- can be converted to an `int`

This means values like `4.` (a float but a whole number) or
`np.array([4])[0]` (i.e., type `numpy.int64` is not `int`)
are now a valid spatial_dimension value.
They all will, however, be converted to `int` as the attribute value
of the resulting `MultiIndexSet` instance.

This commit should resolve Issue #77.

Additional changes:

- A new verification function `verify_spatial_dimension()` is added to
  the module `minterpy.core.verification`. The function checks for correctness
  of spatial dimension and forces the verified value to be of `int` type.
- `lp_degree` in `MultiIndexSet.from_degree()` may now throw a custom error
   message when a `ValueError` is raised.
- A new test suite for the functions in `minterpy.core.verification` is added.
- The test suite on `MultiIndexSet.from_degree` is extended to include
  testing the behavior of passing different types and values
  (both valid and invalid) for spatial dimension and lp-degree.