Refactor subset checking of MultiIndexSet instances.
The method `is_sub_index_set_of()` of the `MultiIndexSet` class
is now renamed to `is_subset()` for brevity and consistency with
Python built-in sets.

The instance of the class now supports also the operator `<=`
for subset checking.
By default via the operator,  if there's a different in spatial
dimension, the set with a lower dimension is expanded.
This behavior can be changed via a parameter in the method call.

This commit should partially resolve Issue #129.

Some additional changes include updating the test suite to include
testing the subset checking.