Add proper subset checking of MultiIndexSet instances.
The method `is_propsubset()` is added for the `MultiIndexSet` class
to carry out proper subset checking.
The instance of the class now also supports the operator `<` for
proper subset checking. By default via operator, if there's a different
in the spatial dimension, the set with a lower dimension is expanded.
This behavior can be changed via the method call.

This commit should partially resolve Issue #129.

Some additional changes include updating the test suite to include
testing for proper subset and the How-To Guide on set operations
involving `MultiIndexSet` instances.