Add proper superset checking of MultiIndexSet instances.
The method `is_propsuperset()` is added for the `MultiIndexSet` class to
carry out proper superset checking.
The instance of the class now also supports the operator `>` for proper
superset checking.
By default via operator, if there's a different in the spatial
dimension, the set with a lower dimension is expanded.
This behavior can be changed via the method call.

This commit should partially resolve Issue #129.

Some additional changes include updating the test suite to include
testing for proper superset checking and the How-To Guide on set
operations involving `MultiIndexSet`.