Add disjoint checking of MultiIndexSet instances.
The method `is_disjoint()` is added for the `MultiIndexSet` class to
carry out disjoint check between two instances of the class.
There is no equivalent operator for this method.
By default, if there is a difference in the spatial dimension, there is
no dimension expansion and an exception is raised.
This behavior can be changed via the `expand_dim` parameter.

This commit should partially resolve Issue #129.

Some additional changes include updating the test suite to include
testing for disjoint check and the How-To Guide on set operations
involving `MultiIndexSet`.