Modify __repr__ and __str__ of MultiIndexSet.
The dunder methods `__repr__()` and `__str__()` of the `MultiIndexSet`
class are revised. Following a Python best-practice, `__repr__()` now
includes the information on lp-degree as the minimum required
information to reproduce the instance.
`__str__()` now includes information regarding the spatial dimension
(`m`), polynomial degree (`n`), and lp-degree (`p`) to be more
user-friendly.

This commit should resolve Issue #133.

Additionally, the test suite is updated with the tests for the string
representation of the `MultiIndexSet` instances (both `repr()` and
`str()`).