Update the module-level docstring of multi_index.py.
The in-code documentation of the module `minterpy.core.multi_index`
(i.e., the module that contains the implementation of the `MultiIndexSet`
class) is updated at the module level. Additionally, several missing
docstrings for some methods are added or updated.

There are no changes in the code functionalities.

This commit should resolve Issue #134.